Move Metadata into Core
Transaction Metadata now resides in core/metadata. The rest of the
metadata stays in chain but now in chain/auxdata. Now any
metadata-defined library won't need to require chain (e.g. cip25, etc)
which should greatly reduce the compiled size when crypto isn't needed.

chain/address had some refactors as I attempted to move it to its own
lib but it has too many dependencies. I kept these changes in case we
try to move it again as now it'll be easier.

Serialize's to_bytes has been reworked as discussed previously.
