+++ chain/src/lib.rs
@@ -7,17 +7,27 @@ use cbor_event::Special as CBORSpecial;
 use std::collections::BTreeMap;
 use std::convert::{From, TryFrom};
 
+pub mod address;
+pub mod auxdata;
 pub mod serialization;
 pub mod crypto;
 
+use address::*;
+use auxdata::*;
 
 use crypto::*;
 
 //pub mod legacy_address;
 
 pub use cardano_multiplatform_lib_core::{
+    Int,
+    Epoch,
+    Slot,
+    TransactionIndex,
+    CertificateIndex,
     ordered_hash_map::OrderedHashMap,
     error::{DeserializeError, DeserializeFailure},
+    metadata::{TransactionMetadatum, TransactionMetadatumLabel},
     serialization::{Serialize, Deserialize, StringEncoding, LenEncoding},
 };
 
@@ -47,18 +57,11 @@ pub type DatumHash = DataHash;
 
 pub type DeltaCoin = Int;
 
-pub type Epoch = u64;
-
-pub type Slot = u64;
-
-pub type CertificateIndex = u64;
-
 pub type Genesishashs = Vec<GenesisHash>;
 
 // TODO: fix cddl-codegen to avoid generating this and make it a direct alias not a declared one
 pub type Int64 = i64;
 
-pub type Metadata = OrderedHashMap<TransactionMetadatumLabel, TransactionMetadatum>;
 
 pub type Mint = OrderedHashMap<PolicyId, OrderedHashMap<AssetName, Int64>>;
 
@@ -91,89 +94,10 @@ pub type ShelleyAuxData = OrderedHashMap<TransactionMetadatumLabel, TransactionM
 
 pub type SubCoin = PositiveInterval;
 
-pub type TransactionIndex = u16;
-
-pub type TransactionMetadatumLabel = u64;
-
 pub type Withdrawals = OrderedHashMap<RewardAccount, Coin>;
 
 // for enums + byron. see comment in Cargo.toml
 use wasm_bindgen::prelude::wasm_bindgen;
-
-#[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema, Derivative)]
-#[derivative(Eq, PartialEq, Ord="feature_allow_slow_enum", PartialOrd="feature_allow_slow_enum", Hash)]
-pub enum Int {
-    Uint {
-        value: u64,
-        #[derivative(PartialEq="ignore", Ord="ignore", PartialOrd="ignore", Hash="ignore")]
-        #[serde(skip)]
-        encoding: Option<cbor_event::Sz>,
-    }
-    ,
-    Nint {
-        value: u64,
-        #[derivative(PartialEq="ignore", Ord="ignore", PartialOrd="ignore", Hash="ignore")]
-        #[serde(skip)]
-        encoding: Option<cbor_event::Sz>,
-    }
-    ,
-}
-
-#[derive(Clone, Debug)]
-pub enum IntError {
-    Bounds(std::num::TryFromIntError),
-    Parsing(std::num::ParseIntError),
-}
-
-impl Int {
-    pub fn new_uint(value: u64) -> Self {
-        Self::Uint {
-            value,
-            encoding: None,
-        }
-    }
-
-    /// * `value` - Value as encoded in CBOR - note: a negative `x` here would be `|x + 1|` due to CBOR's `nint` encoding e.g. to represent -5, pass in 4.
-    pub fn new_nint(value: u64) -> Self {
-        Self::Nint {
-            value,
-            encoding: None,
-        }
-    }
-}
-
-impl std::fmt::Display for Int {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Uint{ value, .. } => write!(f, "{}", value),
-            Self::Nint{ value, .. } => write!(f, "-{}", value + 1),
-        }
-    }
-}
-
-impl std::str::FromStr for Int {
-    type Err = IntError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let x = i128::from_str(s).map_err(IntError::Parsing)?;
-        Self::try_from(x).map_err(IntError::Bounds)
-    }
-}
-
-impl TryFrom<i128> for Int {
-    type Error = std::num::TryFromIntError;
-
-    fn try_from(x: i128) -> Result<Self, Self::Error> {
-        if x >= 0 {
-            u64::try_from(x).map(|x| Self::Uint{ value: x, encoding: None })
-        }
-        else {
-            u64::try_from((x + 1).abs()).map(|x| Self::Nint{ value: x, encoding: None })
-        }
-    }
-}
-
-pub mod address;
 
 pub use address::*;
 
@@ -186,17 +110,6 @@ pub use block::*;
 pub mod certs;
 
 pub use certs::*;
-
-
-//pub mod crypto;
-
-//pub use crypto::*;
-
-
-pub mod metadata;
-
-pub use metadata::*;
-
 
 pub mod plutus;
 
@@ -484,67 +397,6 @@ impl Script2 {
         Self {
             plutus_v2_script,
             encodings: None,
-        }
-    }
-}
-
-#[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema, Derivative)]
-#[derivative(Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub struct KeyStakeCredential {
-    pub addr_keyhash: AddrKeyhash,
-    #[derivative(PartialEq="ignore", Ord="ignore", PartialOrd="ignore", Hash="ignore")]
-    #[serde(skip)]
-    pub encodings: Option<StakeCredentialEncoding>,
-}
-
-impl KeyStakeCredential {
-    pub fn new(addr_keyhash: AddrKeyhash) -> Self {
-        Self {
-            addr_keyhash,
-            encodings: None,
-        }
-    }
-}
-
-#[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema, Derivative)]
-#[derivative(Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub struct ScriptStakeCredential {
-    pub scripthash: ScriptHash,
-    #[derivative(PartialEq="ignore", Ord="ignore", PartialOrd="ignore", Hash="ignore")]
-    #[serde(skip)]
-    pub encodings: Option<StakeCredentialEncoding>,
-}
-
-impl ScriptStakeCredential {
-    pub fn new(scripthash: ScriptHash) -> Self {
-        Self {
-            scripthash,
-            encodings: None,
-        }
-    }
-}
-
-#[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema, Derivative)]
-#[derivative(Eq, PartialEq, Ord="feature_allow_slow_enum", PartialOrd="feature_allow_slow_enum", Hash)]
-pub enum StakeCredential {
-    Key(KeyStakeCredential),
-    Script(ScriptStakeCredential),
-}
-
-impl StakeCredential {
-    pub fn new_key(addr_keyhash: Ed25519KeyHash) -> Self {
-        Self::Key(KeyStakeCredential::new(addr_keyhash))
-    }
-
-    pub fn new_script(scripthash: ScriptHash) -> Self {
-        Self::Script(ScriptStakeCredential::new(scripthash))
-    }
-
-    pub fn to_raw_bytes(&self) -> &[u8] {
-        use cardano_multiplatform_lib_crypto::RawBytesEncoding;
-        match self {
-            Self::Key(key) => key.addr_keyhash.to_raw_bytes(),
-            Self::Script(script) => script.scripthash.to_raw_bytes(),
         }
     }
 }
