+++ chain/src/serialization.rs
@@ -13,20 +13,6 @@ pub use cardano_multiplatform_lib_core::{
     }
 };
 
-impl Serialize for Address {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        let raw_bytes = self.to_raw_bytes();
-        serializer.write_bytes_sz(&raw_bytes, self.encoding().map(|encs| encs.bytes_encoding.clone()).unwrap_or_default().to_str_len_sz(raw_bytes.len() as u64, force_canonical))
-    }
-}
-
-impl Deserialize for Address {
-    fn deserialize<R: BufRead>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let (raw_bytes, encoding) = raw.bytes_sz()?;
-        Self::from_bytes_impl(raw_bytes.as_ref(), Some(encoding.into()))
-    }
-}
-
 impl Serialize for AlonzoAuxData {
     fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
         serializer.write_tag_sz(259u64, fit_sz(259u64, self.encodings.as_ref().map(|encs| encs.tag_encoding).unwrap_or_default(), force_canonical))?;
@@ -1570,27 +1556,6 @@ impl Deserialize for I0OrI1 {
             };
             Err(DeserializeError::new("I0OrI1", DeserializeFailure::NoVariantMatched.into()))
         })().map_err(|e| e.annotate("I0OrI1"))
-    }
-}
-
-impl Serialize for Int {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        match self {
-            Self::Uint{ value, encoding } => serializer.write_unsigned_integer_sz(*value, fit_sz(*value, *encoding, force_canonical)),
-            Self::Nint{ value, encoding } => serializer.write_negative_integer_sz(-((*value as i128) + 1), fit_sz(*value, *encoding, force_canonical)),
-        }
-    }
-}
-
-impl Deserialize for Int {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            match raw.cbor_type()? {
-                cbor_event::Type::UnsignedInteger => raw.unsigned_integer_sz().map(|(x, enc)| Self::Uint{ value: x, encoding: Some(enc) }).map_err(std::convert::Into::into),
-                cbor_event::Type::NegativeInteger => raw.negative_integer_sz().map(|(x, enc)| Self::Nint{ value: (-1 - x) as u64, encoding: Some(enc) }).map_err(std::convert::Into::into),
-                _ => Err(DeserializeFailure::NoVariantMatched.into()),
-            }
-        })().map_err(|e| e.annotate("Int"))
     }
 }
 
@@ -3540,23 +3505,6 @@ impl Deserialize for RequiredSigners {
     }
 }
 
-impl Serialize for RewardAccount {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        self.clone().to_address().serialize(serializer, force_canonical)
-    }
-}
-
-impl Deserialize for RewardAccount {
-    fn deserialize<R: BufRead>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let (raw_bytes, encoding) = raw.bytes_sz()?;
-        match Address::from_bytes_impl(raw_bytes.as_ref(), Some(encoding.into()))? {
-            Address::Reward(reward_address) => Ok(reward_address),
-            // first byte must exist or else from_bytes_impl would have errored
-            _ => Err(DeserializeFailure::BadAddressType(raw_bytes[0]).into())
-        }
-    }
-}
-
 impl Serialize for Script {
     fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
         match self {
@@ -4377,162 +4325,6 @@ impl DeserializeEmbeddedGroup for SingleHostName {
                 len_encoding,
                 index_0_encoding,
                 port_encoding,
-            }),
-        })
-    }
-}
-
-impl Serialize for StakeCredential {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        match self {
-            StakeCredential::Key(stake_credential0) => stake_credential0.serialize(serializer, force_canonical),
-            StakeCredential::Script(stake_credential1) => stake_credential1.serialize(serializer, force_canonical),
-        }
-    }
-}
-
-impl Deserialize for StakeCredential {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array_sz()?;
-            let outer_len_encoding: LenEncoding = len.into();
-            let mut read_len = CBORReadLen::new(len);
-            let initial_position = raw.as_mut_ref().seek(SeekFrom::Current(0)).unwrap();
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                Ok(KeyStakeCredential::deserialize_as_embedded_group(raw, &mut read_len, len)?)
-            })(raw)
-            {
-                Ok(stake_credential0) => return Ok(Self::Key(stake_credential0)),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                Ok(ScriptStakeCredential::deserialize_as_embedded_group(raw, &mut read_len, len)?)
-            })(raw)
-            {
-                Ok(stake_credential1) => return Ok(Self::Script(stake_credential1)),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            match len {
-                cbor_event::LenSz::Len(_, _) => (),
-                cbor_event::LenSz::Indefinite => match raw.special()? {
-                    CBORSpecial::Break => (),
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            Err(DeserializeError::new("StakeCredential", DeserializeFailure::NoVariantMatched.into()))
-        })().map_err(|e| e.annotate("StakeCredential"))
-    }
-}
-
-impl Serialize for KeyStakeCredential {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array_sz(self.encodings.as_ref().map(|encs| encs.len_encoding).unwrap_or_default().to_len_sz(2, force_canonical))?;
-        self.serialize_as_embedded_group(serializer, force_canonical)
-    }
-}
-
-impl SerializeEmbeddedGroup for KeyStakeCredential {
-    fn serialize_as_embedded_group<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_unsigned_integer_sz(0u64, fit_sz(0u64, self.encodings.as_ref().map(|encs| encs.index_0_encoding.clone()).unwrap_or_default(), force_canonical))?;
-        self.addr_keyhash.serialize(serializer, force_canonical)?;
-        self.encodings.as_ref().map(|encs| encs.len_encoding).unwrap_or_default().end(serializer, force_canonical)
-    }
-}
-
-impl Deserialize for KeyStakeCredential {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array_sz()?;
-            let mut read_len = CBORReadLen::new(len);
-            read_len.read_elems(2)?;
-            let ret = Self::deserialize_as_embedded_group(raw, &mut read_len, len);
-            match len {
-                cbor_event::LenSz::Len(_, _) => (),
-                cbor_event::LenSz::Indefinite => match raw.special()? {
-                    CBORSpecial::Break => (),
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            ret
-        })().map_err(|e| e.annotate("KeyStakeCredential"))
-    }
-}
-
-impl DeserializeEmbeddedGroup for KeyStakeCredential {
-    fn deserialize_as_embedded_group<R: BufRead + Seek>(raw: &mut Deserializer<R>, read_len: &mut CBORReadLen, len: cbor_event::LenSz) -> Result<Self, DeserializeError> {
-        let len_encoding = len.into();
-        let index_0_encoding = (|| -> Result<_, DeserializeError> {
-            let (index_0_value, index_0_encoding) = raw.unsigned_integer_sz()?;
-            if index_0_value != 0 {
-                return Err(DeserializeFailure::FixedValueMismatch{ found: Key::Uint(index_0_value), expected: Key::Uint(0) }.into());
-            }
-            Ok(Some(index_0_encoding))
-        })().map_err(|e| e.annotate("index_0"))?;
-        let addr_keyhash = (|| -> Result<_, DeserializeError> {
-            Ok(Ed25519KeyHash::deserialize(raw)?)
-        })().map_err(|e| e.annotate("addr_keyhash"))?;
-        Ok(KeyStakeCredential {
-            addr_keyhash,
-            encodings: Some(StakeCredentialEncoding {
-                len_encoding,
-                index_0_encoding,
-            }),
-        })
-    }
-}
-
-impl Serialize for ScriptStakeCredential {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array_sz(self.encodings.as_ref().map(|encs| encs.len_encoding).unwrap_or_default().to_len_sz(2, force_canonical))?;
-        self.serialize_as_embedded_group(serializer, force_canonical)
-    }
-}
-
-impl SerializeEmbeddedGroup for ScriptStakeCredential {
-    fn serialize_as_embedded_group<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_unsigned_integer_sz(1u64, fit_sz(1u64, self.encodings.as_ref().map(|encs| encs.index_0_encoding.clone()).unwrap_or_default(), force_canonical))?;
-        self.scripthash.serialize(serializer, force_canonical)?;
-        self.encodings.as_ref().map(|encs| encs.len_encoding).unwrap_or_default().end(serializer, force_canonical)
-    }
-}
-
-impl Deserialize for ScriptStakeCredential {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array_sz()?;
-            let mut read_len = CBORReadLen::new(len);
-            read_len.read_elems(2)?;
-            let ret = Self::deserialize_as_embedded_group(raw, &mut read_len, len);
-            match len {
-                cbor_event::LenSz::Len(_, _) => (),
-                cbor_event::LenSz::Indefinite => match raw.special()? {
-                    CBORSpecial::Break => (),
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            ret
-        })().map_err(|e| e.annotate("ScriptStakeCredential"))
-    }
-}
-
-impl DeserializeEmbeddedGroup for ScriptStakeCredential {
-    fn deserialize_as_embedded_group<R: BufRead + Seek>(raw: &mut Deserializer<R>, read_len: &mut CBORReadLen, len: cbor_event::LenSz) -> Result<Self, DeserializeError> {
-        let len_encoding = len.into();
-        let index_0_encoding = (|| -> Result<_, DeserializeError> {
-            let (index_0_value, index_0_encoding) = raw.unsigned_integer_sz()?;
-            if index_0_value != 1 {
-                return Err(DeserializeFailure::FixedValueMismatch{ found: Key::Uint(index_0_value), expected: Key::Uint(1) }.into());
-            }
-            Ok(Some(index_0_encoding))
-        })().map_err(|e| e.annotate("index_0"))?;
-        let scripthash = (|| -> Result<_, DeserializeError> {
-            Ok(ScriptHash::deserialize(raw)?)
-        })().map_err(|e| e.annotate("scripthash"))?;
-        Ok(ScriptStakeCredential {
-            scripthash,
-            encodings: Some(StakeCredentialEncoding {
-                len_encoding,
-                index_0_encoding,
             }),
         })
     }
@@ -5449,130 +5241,6 @@ impl Deserialize for TransactionInput {
                 }),
             })
         })().map_err(|e| e.annotate("TransactionInput"))
-    }
-}
-
-impl Serialize for TransactionMetadatum {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>, force_canonical: bool) -> cbor_event::Result<&'se mut Serializer<W>> {
-        match self {
-            TransactionMetadatum::MapTransactionMetadatumToTransactionMetadatum{ map_transaction_metadatum_to_transaction_metadatum, map_transaction_metadatum_to_transaction_metadatum_encoding } => {
-                serializer.write_map_sz(map_transaction_metadatum_to_transaction_metadatum_encoding.to_len_sz(map_transaction_metadatum_to_transaction_metadatum.len() as u64, force_canonical))?;
-                let mut key_order = map_transaction_metadatum_to_transaction_metadatum.iter().map(|(k, v)| {
-                    let mut buf = cbor_event::se::Serializer::new_vec();
-                    k.serialize(&mut buf, force_canonical)?;
-                    Ok((buf.finalize(), k, v))
-                }).collect::<Result<Vec<(Vec<u8>, &_, &_)>, cbor_event::Error>>()?;
-                if force_canonical {
-                    key_order.sort_by(|(lhs_bytes, _, _), (rhs_bytes, _, _)| {
-                        match lhs_bytes.len().cmp(&rhs_bytes.len()) {
-                            std::cmp::Ordering::Equal => lhs_bytes.cmp(&rhs_bytes),
-                            diff_ord => diff_ord,
-                        }
-                    });
-                }
-                for (key_bytes, key, value) in key_order {
-                    serializer.write_raw_bytes(&key_bytes)?;
-                    value.serialize(serializer, force_canonical)?;
-                }
-                map_transaction_metadatum_to_transaction_metadatum_encoding.end(serializer, force_canonical)
-            },
-            TransactionMetadatum::ArrTransactionMetadatum{ arr_transaction_metadatum, arr_transaction_metadatum_encoding } => {
-                serializer.write_array_sz(arr_transaction_metadatum_encoding.to_len_sz(arr_transaction_metadatum.len() as u64, force_canonical))?;
-                for element in arr_transaction_metadatum.iter() {
-                    element.serialize(serializer, force_canonical)?;
-                }
-                arr_transaction_metadatum_encoding.end(serializer, force_canonical)
-            },
-            TransactionMetadatum::Int(int) => {
-                int.serialize(serializer, force_canonical)
-            },
-            TransactionMetadatum::Bytes{ bytes, bytes_encoding } => {
-                serializer.write_bytes_sz(&bytes, bytes_encoding.to_str_len_sz(bytes.len() as u64, force_canonical))
-            },
-            TransactionMetadatum::Text{ text, text_encoding } => {
-                serializer.write_text_sz(&text, text_encoding.to_str_len_sz(text.len() as u64, force_canonical))
-            },
-        }
-    }
-}
-
-impl Deserialize for TransactionMetadatum {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let initial_position = raw.as_mut_ref().seek(SeekFrom::Current(0)).unwrap();
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                let mut map_transaction_metadatum_to_transaction_metadatum_table = OrderedHashMap::new();
-                let map_transaction_metadatum_to_transaction_metadatum_len = raw.map_sz()?;
-                let map_transaction_metadatum_to_transaction_metadatum_encoding = map_transaction_metadatum_to_transaction_metadatum_len.into();
-                while match map_transaction_metadatum_to_transaction_metadatum_len { cbor_event::LenSz::Len(n, _) => map_transaction_metadatum_to_transaction_metadatum_table.len() < n as usize, cbor_event::LenSz::Indefinite => true, } {
-                    if raw.cbor_type()? == CBORType::Special {
-                        assert_eq!(raw.special()?, CBORSpecial::Break);
-                        break;
-                    }
-                    let map_transaction_metadatum_to_transaction_metadatum_key = TransactionMetadatum::deserialize(raw)?;
-                    let map_transaction_metadatum_to_transaction_metadatum_value = TransactionMetadatum::deserialize(raw)?;
-                    if map_transaction_metadatum_to_transaction_metadatum_table.insert(map_transaction_metadatum_to_transaction_metadatum_key.clone(), map_transaction_metadatum_to_transaction_metadatum_value).is_some() {
-                        return Err(DeserializeFailure::DuplicateKey(Key::Str(String::from("some complicated/unsupported type"))).into());
-                    }
-                }
-                Ok((map_transaction_metadatum_to_transaction_metadatum_table, map_transaction_metadatum_to_transaction_metadatum_encoding))
-            })(raw)
-            {
-                Ok((map_transaction_metadatum_to_transaction_metadatum, map_transaction_metadatum_to_transaction_metadatum_encoding)) => return Ok(Self::MapTransactionMetadatumToTransactionMetadatum {
-                    map_transaction_metadatum_to_transaction_metadatum,
-                    map_transaction_metadatum_to_transaction_metadatum_encoding,
-                }),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                let mut arr_transaction_metadatum_arr = Vec::new();
-                let len = raw.array_sz()?;
-                let arr_transaction_metadatum_encoding = len.into();
-                while match len { cbor_event::LenSz::Len(n, _) => arr_transaction_metadatum_arr.len() < n as usize, cbor_event::LenSz::Indefinite => true, } {
-                    if raw.cbor_type()? == CBORType::Special {
-                        assert_eq!(raw.special()?, CBORSpecial::Break);
-                        break;
-                    }
-                    arr_transaction_metadatum_arr.push(TransactionMetadatum::deserialize(raw)?);
-                }
-                Ok((arr_transaction_metadatum_arr, arr_transaction_metadatum_encoding))
-            })(raw)
-            {
-                Ok((arr_transaction_metadatum, arr_transaction_metadatum_encoding)) => return Ok(Self::ArrTransactionMetadatum {
-                    arr_transaction_metadatum,
-                    arr_transaction_metadatum_encoding,
-                }),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                Ok(Int::deserialize(raw)?)
-            })(raw)
-            {
-                Ok(int) => return Ok(Self::Int(int)),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                Ok(raw.bytes_sz().map(|(bytes, enc)| (bytes, StringEncoding::from(enc)))?)
-            })(raw)
-            {
-                Ok((bytes, bytes_encoding)) => return Ok(Self::Bytes {
-                    bytes,
-                    bytes_encoding,
-                }),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-                Ok(raw.text_sz().map(|(s, enc)| (s, StringEncoding::from(enc)))?)
-            })(raw)
-            {
-                Ok((text, text_encoding)) => return Ok(Self::Text {
-                    text,
-                    text_encoding,
-                }),
-                Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-            };
-            Err(DeserializeError::new("TransactionMetadatum", DeserializeFailure::NoVariantMatched.into()))
-        })().map_err(|e| e.annotate("TransactionMetadatum"))
     }
 }
